@@ -10,12 +10,11 @@
         }
     ],
     "require": {
-        "php": "^7.3 || ~8.0.0 || ~8.1.0",
-        "phpunit/phpunit": "^9.0"
+        "php": "^8.1",
+        "phpunit/phpunit": "^10.0"
     },
-    "conflict": {
-        "phpunit/phpunit": "9.5.0"
-    },
+    "minimum-stability": "dev",
+    "prefer-stable": true,
     "autoload": {
         "psr-4": {
             "Zalas\\PHPUnit\\Globals\\": "src/"
@@ -28,7 +27,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.x-dev"
+            "dev-master": "3.x-dev"
         }
     }
 }
