@@ -3,13 +3,12 @@ declare(strict_types=1);
 
 namespace Zalas\PHPUnit\Globals\Tests;
 
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
-class AnnotationExtensionWithDataProviderTest extends TestCase
+class AttributeExtensionWithDataProviderTest extends TestCase
 {
-    /**
-     * @dataProvider provider
-     */
+    #[DataProvider('provider')]
     public function test_it_handles_dataproviders()
     {
         $this->assertTrue(true, 'It lets the test cases run normally');
